fix(dashboard): move title parser
