+++ src/website/client/src/utils/parsers.ts
@@ -53,3 +53,17 @@ function editDistance(s1: string, s2: string) {
     }
     return costs[s2.length];
 }
+
+export function titleCase(str: string) {
+    if (str === "nsfw") {
+        return "NSFW";
+    }
+    const splitStr = str.toLowerCase().split(' ');
+    for (let i = 0; i < splitStr.length; i++) {
+        // You do not need to check if i is larger than splitStr length, as your for does that for you
+        // Assign it back to the array
+        splitStr[i] = splitStr[i].charAt(0).toUpperCase() + splitStr[i].substring(1);
+    }
+    // Directly return the joined string
+    return splitStr.join(' ');
+}
